Conflicts:
	spatialops/FieldExpressions.fmr
	spatialops/FieldExpressions.h
	spatialops/FieldExpressionsFDStencil2.fmr
	spatialops/FieldExpressionsFDStencil2.h
	spatialops/FieldExpressionsStencil2.fmr
	spatialops/FieldExpressionsStencil2.h
	spatialops/FieldExpressionsStencil4.fmr
	spatialops/FieldExpressionsStencil4.h
	test/StencilTest2.cpp
	test/StencilTest4.cpp
	test/StencilTest5.cpp
	test/StencilTest6.cpp
	test/StencilTest7.cpp
	test/StencilTest8.cpp
	test/StencilTest9.cpp

+++ spatialops/FieldExpressions.h
@@ -5928,7 +5928,7 @@
                                                                            NeboExpression<ExprType,
                                                                                           FieldType>
                                                                            const & initial_rhs,
-                                                                           int const
+                                                                           size_t const
                                                                            number_of_partitions) {
 
              typename NeboField<Initial, FieldType>::ResizePrepType typedef LhsType;
